requirements.txt
biopython==1.81
numpy==1.24.3
pandas==2.0.2
pypdb==2.2
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2